Fix Windows install for gfortran

+++ recipes/gfortran/all/conanfile.py
@@ -34,13 +34,12 @@ class GFortranConan(ConanFile):
     def source(self):
         url = self.conan_data["sources"][self.version]["url"]
         for it in url.keys():
-            if self.settings.os == "Windows":
-                if it == "Windows":
-                    filename = url[it]["filename"]
-                    tools.download(**url[it])
-                    self.run("7z x {0}".format(filename))
-                    os.unlink(filename)
-                    os.rename("mingw64", "source_subfolder_Windows")
+            if self.settings.os == "Windows" and it == "Windows":
+                filename = url[it]["filename"]
+                tools.download(**url[it])
+                self.run("7z x {0}".format(filename))
+                os.unlink(filename)
+                os.rename("mingw64", "source_subfolder_Windows")
             else:
                 tools.get(**url[it])
                 pattern = "gcc-*" if it == "Linux" else "usr"
